+++ playwright/utils/fetchOmnichannelConfig.ts
@@ -1,23 +1,4 @@
-import fs from 'fs';
-import path from 'path';
-import YAML from 'yaml';
-
-const fetchTestConfig = () => {
-    const testConfigFilePath = path.join(path.dirname(__dirname), 'test.config.yml');
-    let testConfig = null;
-    try {
-        const fileData = fs.readFileSync(testConfigFilePath, 'utf8');
-        testConfig = YAML.parse(fileData);
-    } catch {
-        throw new Error(`Unable to process test config file ${testConfigFilePath}`);
-    }
-
-    if (!testConfig) {
-        throw new Error('Test config file is empty');
-    }
-
-    return testConfig;
-};
+import fetchTestConfig from "./fetchTestConfig";
 
 const fetchOmnichannelConfig = (scenario = ""): object => {
     const testConfig = fetchTestConfig();
